requirements.txt
Django~=5.2.1
django-extensions~=4.1
djangorestframework~=3.16.0
gunicorn~=23.0.0
psycopg[binary]~=3.2.9
